feat: Add ActivityIcon component and integrate sidebar context for dynamic footer management
