+++ frontend/src/components/TimelineSidebarCalendar.tsx
@@ -1,14 +1,22 @@
 import { Calendar } from "@/components/ui/calendar"
+import { useSidebar } from "@/components/ui/sidebar"
+import { cn } from "@/lib/utils"
 
 export function TimelineSidebarCalendar({ date, onDateChange }: { date: Date, onDateChange: (date: Date) => void }) {
+  const { state } = useSidebar()
+
+  if (state === "collapsed") {
+    return null
+  }
+
   return (
-    <div className="p-4">
+    <div className={cn("flex justify-center py-4", state === "expanded" && "animate-in fade-in")}>
       <Calendar
         mode="single"
         selected={date}
-        onSelect={onDateChange}
+        onSelect={(d) => d && onDateChange(d)}
         required
-        className="rounded-md border shadow"
+        className="rounded-md"
       />
     </div>
   )
